Move common settings Composables to settings/mobile Possibly move to a new common UI module in the future
