+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/dndsync/ui/DnDSyncSettingsScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import com.boswelja.smartwatchextensions.R
 import com.boswelja.smartwatchextensions.common.ui.Card
 import com.boswelja.smartwatchextensions.common.ui.CardHeader
-import com.boswelja.smartwatchextensions.common.ui.SwitchSetting
+import com.boswelja.smartwatchextensions.settings.ui.SwitchSetting
 import com.boswelja.smartwatchextensions.settings.BoolSettingKeys.DND_SYNC_TO_PHONE_KEY
 import com.boswelja.smartwatchextensions.settings.BoolSettingKeys.DND_SYNC_WITH_THEATER_KEY
 import kotlinx.coroutines.Dispatchers
