+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/proximity/ui/ProximitySettingsScreen.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.unit.dp
 import com.boswelja.smartwatchextensions.R
 import com.boswelja.smartwatchextensions.common.ui.Card
 import com.boswelja.smartwatchextensions.common.ui.CardHeader
-import com.boswelja.smartwatchextensions.common.ui.SwitchSetting
+import com.boswelja.smartwatchextensions.settings.ui.SwitchSetting
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.getViewModel
