+++ settings/mobile/build.gradle.kts
@@ -24,6 +24,7 @@ kotlin {
         val androidMain by getting {
             dependencies {
                 implementation(libs.sqldelight.android)
+                implementation(libs.bundles.compose.mobile)
             }
         }
         val androidTest by getting {
@@ -34,8 +35,36 @@ kotlin {
     }
 }
 
+android {
+    buildFeatures.compose = true
+    composeOptions {
+        kotlinCompilerExtensionVersion = libs.versions.composeCompiler.get()
+    }
+}
+
 sqldelight {
     database("WatchSettingsDatabase") {
         packageName = "com.boswelja.smartwatchextensions.settings.database"
+    }
+}
+
+// Workaround for https://youtrack.jetbrains.com/issue/KT-38694
+configurations {
+    create("composeCompiler") {
+        isCanBeConsumed = false
+    }
+}
+dependencies {
+    "composeCompiler"("androidx.compose.compiler:compiler:${libs.versions.composeCompiler.get()}")
+}
+android {
+    afterEvaluate {
+        val composeCompilerJar = configurations["composeCompiler"]
+            .resolve()
+            .singleOrNull()
+            ?: error("Missing Compose compiler")
+        tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+            kotlinOptions.freeCompilerArgs += listOf("-Xuse-ir", "-Xplugin=$composeCompilerJar")
+        }
     }
 }
